Integrated qp route to homepage
